requirements.txt
fastapi==0.115.11
uvicorn==0.24.0.post1
python-multipart==0.0.6
psycopg2-binary==2.9.10
